Enforce at_least_one_of on URL map match rules
This matches the description of those associated fields. Without going into detail, without this enforcement, Google silently drops routing rules - very bad. While that's a Google Cloud problem, since the docs themselves state this requirement, my thinking is why not have the tools that are generated from this repo catch this early on as well.

Ref: https://cloud.google.com/compute/docs/reference/rest/v1/urlMaps (search `pathMatchers[].routeRules[].matchRules[]`)

> Only one of prefixMatch, fullPathMatch or regexMatch must be specified.